requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Babel==2.13.1
bcrypt==4.1.3
certifi==2024.6.2
click==8.1.7
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.5
redis-dict==2.5.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
starlette-admin==0.14.0
Starlette-Login==0.2.1
typing_extensions==4.12.2
uvicorn==0.30.1
yarl==1.9.4
